chore: update changelog for 0.12.0 release (#665)
Note that release 0.12.0 will not be tagged on this commit. But on `6c08a05254da556a518dcd03f536e2251b0656d6`. This is because at the time of the release there are new features merged in `main` for the new CRD/operator that we don't want to release yet. But interspersed in these changes are improvements and bug fixes for the old operator that we want to release. Any code the for the new operator or CRD included here will not run at all with the old Amalthea helm chart.