feat: known_hosts support initial implementation
In order to clone a git repository using the ssh protocol,
one must accept to connect to the remote service as must
be done when using ssh alone. Since it's an interactive
operation that is not possible when in an init container,
one solution is to provide the known_hosts files through
a ConfigMap containing a list of validated hosts.