feat: quieter idleness (#747)
* refactor(metrics): add a delay before requesting metrics

Metrics generation does not start at pod creation. There's
a minimal time of 10 seconds for accurate data after the
first container is started hence starting requesting early
will just trigger a lot of pod not found error. Introducing
a delay before starting that request reduces the load on the
metrics server as well as lower the amount of logs for what
can be considered as false positive.

* refactor(metrics): only log idleness if the IdleSince is zero

Otherwise, as long as the pod is considered idle, it will
log this message which will be quite a lot depending on the
number of time the reconciliation loop runs per second.

---------

Co-authored-by: Ralf Grubenmann <ralf.grubenmann@sdsc.ethz.ch>