feat: mount project and home (#1023)
Mount scratch, project and home directories based on the system response from the FirecREST API.

Also, attempt to save and rescue the session if the container is killed and restarted. Session recovery may not be successful, as killing the remote session controller may result in the remote job being cancelled. Though this should help recover the session if the remote session controller goes out of memory.

---------

Co-authored-by: Salim Kayal <salim.kayal@idiap.ch>