feat: Add `lastInteraction` property storing last seen user interaction (#1046)
For detecting when to hibernate a session, we use cpu threshold to detect an idle session. If a user explicitly proves to interact with the session, this timestamp is now stored as well. When it exists and it is after the `idleSince` time, it is used as the starting point for counting down the `MaxIdleTime`.

Removes the logic using last request time for detecting idleness.

Makes the previously hard-coded cpu threshold for idleness-checking configurable.

---------

Co-authored-by: Tasko Olevski <olevski90@gmail.com>
Co-authored-by: Flora Thiebaut <flora.thiebaut@sdsc.ethz.ch>